// frontend/src/api/generated/models/saved-search-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * chainlink-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SavedSearchSaveJson } from './saved-search-save-json';
import {
    SavedSearchSaveJsonFromJSON,
    SavedSearchSaveJsonFromJSONTyped,
    SavedSearchSaveJsonToJSON,
    SavedSearchSaveJsonToJSONTyped,
} from './saved-search-save-json';
import type { EntityInfoJson } from './entity-info-json';
import {
    EntityInfoJsonFromJSON,
    EntityInfoJsonFromJSONTyped,
    EntityInfoJsonToJSON,
    EntityInfoJsonToJSONTyped,
} from './entity-info-json';

/**
 * 
 * @export
 * @interface SavedSearchJson
 */
export interface SavedSearchJson {
    /**
     * 
     * @type {string}
     * @memberof SavedSearchJson
     */
    id: string;
    /**
     * 
     * @type {EntityInfoJson}
     * @memberof SavedSearchJson
     */
    entityInfo: EntityInfoJson;
    /**
     * 
     * @type {SavedSearchSaveJson}
     * @memberof SavedSearchJson
     */
    data: SavedSearchSaveJson;
}

/**
 * Check if a given object implements the SavedSearchJson interface.
 */
export function instanceOfSavedSearchJson(value: object): value is SavedSearchJson {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('entityInfo' in value) || value['entityInfo'] === undefined) return false;
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function SavedSearchJsonFromJSON(json: any): SavedSearchJson {
    return SavedSearchJsonFromJSONTyped(json, false);
}

export function SavedSearchJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): SavedSearchJson {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'entityInfo': EntityInfoJsonFromJSON(json['entityInfo']),
        'data': SavedSearchSaveJsonFromJSON(json['data']),
    };
}

export function SavedSearchJsonToJSON(json: any): SavedSearchJson {
    return SavedSearchJsonToJSONTyped(json, false);
}

export function SavedSearchJsonToJSONTyped(value?: SavedSearchJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'entityInfo': EntityInfoJsonToJSON(value['entityInfo']),
        'data': SavedSearchSaveJsonToJSON(value['data']),
    };
}